Cleaned up interface, adding library start/stop
I greatly simplified the creating of new native wrappers by making all
master class members public, allowing them to be called from the Java wrapping
methods.

I also added support for starting/stopping the OpenSL ES library. This is
because OpenSL holds a wakelock when running. Stopping it allows us to release
the wakelock and be a good Android citizen